+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+Feature: 
+
+- `Alert-box` Component : Add the `Dark` status type
+- `Alert-box` Stories and Tests : Update
+
 ## [14.1.0] - 2023-07-31
 
 Feature: